% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_examples.R
\name{add_examples}
\alias{add_examples}
\title{Add examples to the project}
\usage{
add_examples(path, desc, ...)
}
\arguments{
\item{path}{the path to the project folder}

\item{desc}{optional character vector with description for project.}

\item{...}{currently ignored.}
}
\value{
NULL invisibly
}
\description{
This function adds an example mainscript, function and data to
the project so it becomes clear how one could use it.
}
