% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{add_info}
\alias{add_info}
\title{Add an info file to the project}
\usage{
add_info(path, desc)
}
\arguments{
\item{path}{The path to the main folder of the project}

\item{desc}{The name of the project}
}
\value{
NULL invisibly
}
\description{
Function to create a file INFO.md.
}
