% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{add_readme}
\alias{add_readme}
\title{Add a readme file to the project}
\usage{
add_readme(path, desc)
}
\arguments{
\item{path}{The path to the main folder of the project}

\item{desc}{The name of the project}
}
\value{
NULL invisibly
}
\description{
Function to create a file README.md.
}
