% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dirs.R
\name{setup_dirs}
\alias{setup_dirs}
\title{set up the directory structure}
\usage{
setup_dirs(path, structure = c("Data", "Data_Processed", "Functions",
  "Graphs", "Reports"), recursive = TRUE)
}
\arguments{
\item{path}{The path is the main path where the directory structure
will be created.}

\item{structure}{a character vector with the relative paths that have
to be created.}

\item{recursive}{logical. Should elements of any path other than the
last be created? See also \code{\link[base]{dir.create}}}
}
\value{
a logical vector (invisibly) indicating whether the creation
of the directory worked.
}
\description{
Set up a directory structure for a standard data analyis project.
}
\examples{
\dontrun{
setup_dirs(getwd(), c("Data/raw","Data/processed","Output"))
}

}
